refactor(qwen3_5): share the GDN sidecar capture and MTP acceptance gate between dense and MoE
